# public/openapi.yaml
openapi: 3.0.3
info:
  title: EXOS-Tool API
  description: >
    API REST para subir archivos "show tech-support all" de Extreme Networks EXOS
    y consultar análisis, hallazgos y métricas. Autenticación mediante token
    Bearer (Laravel Sanctum, generado en el perfil del usuario → API Tokens).
  version: "1.0"
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /captures/{id}:
    get:
      summary: Detalle de una captura (incluye resumen parseado y advertencias)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Captura con summary }
        "404": { description: No encontrada }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer